Integrate compound property name tests into existing coverage
