+++ tests/main/jsonschema/test_main_jsonschema.py
@@ -61,10 +61,12 @@ from datamodel_code_generator import (
     cached_path_exists,
     chdir,
     generate,
+    generate_dynamic_models,
     load_data,
     load_data_from_path,
 )
 from datamodel_code_generator.__main__ import Exit
+from datamodel_code_generator.config import GenerateConfig
 from datamodel_code_generator.format import Formatter, is_supported_in_black
 from datamodel_code_generator.model import base as model_base
 from datamodel_code_generator.model import get_data_model_types
@@ -125,6 +127,7 @@ from tests.main.conftest import (
     run_main_with_system_exit,
 )
 from tests.main.jsonschema.conftest import EXPECTED_JSON_SCHEMA_PATH, assert_file_content
+from tests.main.payload_validation.constants import COMPOUND_PROPERTY_NAMES_DIAGNOSTIC_CASES
 
 if TYPE_CHECKING:
     from datamodel_code_generator.parser.base import Result
@@ -23385,3 +23388,110 @@ def test_numeric_null_custom_template(output_file: Path, entrypoint: str) -> Non
         invalid_json="2",
         expected_error_type="none_required",
     )
+
+
+COMPOUND_PROPERTY_INPUTS = JSON_SCHEMA_DATA_PATH / "compound_property_names"
+
+
+COMPOUND_PROPERTY_PAYLOADS = DATA_PATH / "payloads" / "compound_property_names"
+
+
+COMPOUND_PROPERTY_EXPECTED = EXPECTED_JSON_SCHEMA_PATH / "compound_property_names"
+
+
+COMPOUND_PROPERTY_CASES = json.loads((COMPOUND_PROPERTY_PAYLOADS / "cases.json").read_text())
+
+
+@pytest.mark.parametrize("name", COMPOUND_PROPERTY_CASES)
+@pytest.mark.parametrize("constraints", [False, True])
+@pytest.mark.parametrize("entry", ["cli", "api", "dynamic"])
+def test_compound_property_name_generation(name: str, constraints: bool, entry: str, output_file: Path) -> None:
+    """Preserve string keys, native acceptance, and deterministic generated output."""
+    input_path = COMPOUND_PROPERTY_INPUTS / f"{name}.json"
+    suffix = (
+        "_legacy_pydantic"
+        if PYDANTIC_V2_ROOT_MODEL_DICT_KEY_FORWARD_REF_NEEDS_SORTING and name in {"enum_refs", "ref_then_any"}
+        else ""
+    )
+    expected_file = f"compound_property_names/{name}_{int(constraints)}{suffix}.py"
+    if entry == "cli":
+        run_main_and_assert(
+            input_path=input_path,
+            output_path=output_file,
+            input_file_type="jsonschema",
+            extra_args=[
+                "--custom-file-header",
+                "# Compound property names",
+                *(["--field-constraints"] if constraints else []),
+            ],
+            assert_func=assert_file_content,
+            expected_file=expected_file,
+        )
+    elif entry == "api":
+        run_generate_file_and_assert(
+            input_path=input_path,
+            output_path=output_file,
+            input_file_type=InputFileType.JsonSchema,
+            custom_file_header="# Compound property names",
+            field_constraints=constraints,
+            assert_func=assert_file_content,
+            expected_file=expected_file,
+        )
+    schema = json.loads(input_path.read_text())
+    Draft202012Validator.check_schema(schema)
+    validator = Draft202012Validator(schema)
+    payloads = json.loads((COMPOUND_PROPERTY_PAYLOADS / f"{name}.json").read_text())
+    assert_output(
+        json.dumps([validator.is_valid(value) for value in payloads], indent=2) + "\n",
+        COMPOUND_PROPERTY_EXPECTED / f"{name}_runtime.txt",
+    )
+    actual = []
+    context = (
+        nullcontext(
+            generate_dynamic_models(schema, config=GenerateConfig(field_constraints=constraints), cache_size=0)["Root"]
+        )
+        if entry == "dynamic"
+        else _generated_model(output_file, "compound_keys", "Root")
+    )
+    with context as model:
+        for value in payloads:
+            if validator.is_valid(value):
+                model.model_validate(value)
+                actual.append(True)
+            else:
+                with pytest.raises(ValidationError):
+                    model.model_validate(value)
+                actual.append(False)
+    assert_output(json.dumps(actual, indent=2) + "\n", COMPOUND_PROPERTY_EXPECTED / f"{name}_runtime.txt")
+
+
+@pytest.mark.parametrize(("case_id", "expected_file"), COMPOUND_PROPERTY_NAMES_DIAGNOSTIC_CASES.items())
+@pytest.mark.parametrize("constraints", [False, True])
+@pytest.mark.parametrize("entry", ["cli", "api"])
+def test_unrepresentable_compound_property_names(
+    case_id: str,
+    expected_file: str,
+    constraints: bool,
+    entry: str,
+    output_file: Path,
+    capsys: pytest.CaptureFixture[str],
+) -> None:
+    """Diagnose unsupported key intersections instead of emitting invalid models."""
+    input_path = DATA_PATH / case_id
+    expected = EXPECTED_JSON_SCHEMA_PATH / expected_file
+    Draft202012Validator.check_schema(json.loads(input_path.read_text(encoding="utf-8")))
+    if entry == "cli":
+        run_main_and_assert(
+            input_path=input_path,
+            output_path=output_file,
+            input_file_type="jsonschema",
+            extra_args=["--field-constraints"] if constraints else [],
+            expected_exit=Exit.ERROR,
+            output_should_not_exist=True,
+            capsys=capsys,
+            expected_stderr_contains=expected.read_text().strip(),
+        )
+    else:
+        with pytest.raises(SchemaParseError) as error:
+            generate(input_path, input_file_type=InputFileType.JsonSchema, field_constraints=constraints)
+        assert_output(str(error.value) + "\n", expected)
